[Fix] 制御系のミスを修正

+++ main/include/adc.hpp
@@ -8,7 +8,7 @@
 
 void initADC();
 
-int BatteryVoltage();
+float BatteryVoltage();
 void WallSensor();
 
 #endif
